requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==1.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
bottle==0.12.19
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
ctop==1.0.0
cycler==0.11.0
datasets==3.2.0
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-pycreds==0.4.0
einops==0.8.0
entrypoints==0.4
exceptiongroup==1.2.2
filelock==3.6.0
flake8==4.0.1
flash_attn==2.7.4.post1
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.29.1
fpdf==1.7.2
frozenlist==1.5.0
fs==2.4.12
fsspec==2024.3.1
future==0.18.2
gast==0.5.2
gitdb==4.0.12
GitPython==3.1.44
Glances==3.2.4.2
google-pasta==0.2.0
grpcio==1.30.2
h11==0.14.0
h5py==3.6.0
h5py.-debian-h5py-serial==3.6.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.28.1
hyperlink==21.0.0
icdiff==2.0.4
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
influxdb==5.3.1
iotop==0.6
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.5
jiter==0.8.2
joblib==0.17.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
kaptan==0.5.12
keras==3.6.0
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libtmux==0.10.1
livereload==2.6.3
lxml==4.8.0
lz4==3.1.3+dfsg
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mkdocs==1.1.2
ml-dtypes==0.5.0
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.4
numpy==1.21.5
nvidia-ml-py==12.555.43
oauthlib==3.2.0
olefile==0.46
openai==1.61.0
opt-einsum==3.3.0
optree==0.13.1
packaging==21.3
pandas==1.3.5
parso==0.8.1
peft==0.14.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipx==1.0.0
platformdirs==2.5.1
ply==3.11
prompt-toolkit==3.0.28
propcache==0.2.1
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.11.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pysmi==0.3.2
pysnmp==4.4.12
pystache==0.6.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-magic==0.4.24
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2024.11.6
requests==2.32.3
rich==11.2.0
safetensors==0.5.2
scikit-learn==0.23.2
scipy==1.8.0
SecretStorage==3.3.1
sentry-sdk==2.20.0
service-identity==18.1.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sos==4.7.2
soupsieve==2.3.1
ssh-import-id==5.11
sympy==1.12
tensorboard==2.18.0
tensorflow==2.18.0
termcolor==1.1.0
tf_keras==2.18.0
threadpoolctl==3.1.0
tmuxp==1.9.2
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tornado==6.1
tqdm==4.67.1
traitlets==5.1.1
transformers==4.48.2
triton==3.1.0
trl==0.14.0
Twisted==22.1.0
typing_extensions==4.12.2
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.3.0
userpath==1.8.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wandb==0.19.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.13.3
xxhash==3.5.0
yarl==1.18.3
zipp==1.0.0
zope.interface==5.4.0
